Move flattenSubOrgs() utility to separate module
